requirements.txt
Flask
mysql-connector-python
oauth2client
gspread
pandas
numpy
matplotlib
pathlib
